Added validators for the commands as well as modified most of the commands to receive the userId so as to perform authorization checks e.g. is the user the creator of whatever he is trying to modify. Further worked on the voting implementation. Added additional properties to most of the Vms for additional info that then got displayed in the front end. Worked on the front end. Added logo.

+++ src/Core/DiscussionBoard.Application/Comments/Commands/DeleteComment/DeleteCommentCommandHandler.cs
@@ -2,6 +2,7 @@
 using DiscussionBoard.Domain.Entities;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -21,6 +22,16 @@ namespace DiscussionBoard.Application.Comments.Commands.DeleteComment
             var comment = await _commentsRepository
                 .All()
                 .SingleOrDefaultAsync(c => c.Id == request.Id);
+
+            if (comment == null)
+            {
+                throw new Exception("Not Found");
+            }
+
+            if (comment.CreatorId != request.CreatorId)
+            {
+                throw new Exception("Unauthorized");
+            }
 
             _commentsRepository.Delete(comment);
             await _commentsRepository.SaveChangesAsync();
